requirements.txt
cachelib==0.9.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dash==2.7.1
dash-bootstrap-components==1.3.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
Flask-Caching==2.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.1
pandas==1.5.2
plotly==5.11.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
requests==2.28.1
sgex==0.5.5
six==1.16.0
tenacity==8.1.0
urllib3==1.26.13
Werkzeug==2.2.2
